requirements.txt
ipython==8.15.0
matplotlib==3.5.1
natsort==8.4.0
nbconvert==7.16.0
nbformat==5.9.2
neptune==1.8.6
numpy==1.21.5
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
pandas==1.4.4
papermill==2.5.0
perlin_noise==1.12
Pillow==10.0.0
rich==13.7.0
scikit_image==0.18.3
scikit_learn==0.24.2
scipy==1.9.1
seaborn==0.11.2
statsmodels==0.13.2
sty==1.0.4
toml==0.10.2
torch==1.13.0
torchvision==0.14.0
tqdm==4.62.3
